Extract ShortcutsList to separate file
