+++ shortcuts-raycast-extension/src/app-shortcuts.tsx
@@ -1,21 +1,9 @@
-import {
-  Action,
-  ActionPanel,
-  closeMainWindow,
-  getFrontmostApplication,
-  getPreferenceValues,
-  Icon,
-  List,
-  PopToRootType,
-} from "@raycast/api";
+import { closeMainWindow, getFrontmostApplication, PopToRootType } from "@raycast/api";
 import { showFailureToast, usePromise } from "@raycast/utils";
 import { useEffect, useState } from "react";
-import { runShortcuts } from "./engine/shortcut-runner";
-import { Application, AtomicShortcut, Keymap, Section } from "./model/internal/internal-models";
+import { Application, Keymap, Section } from "./model/internal/internal-models";
 import useAllShortcuts from "./load/shortcuts-provider";
-import useKeyCodes from "./load/key-codes-provider";
-import { generateHotkeyText } from "./view/hotkey-text-formatter";
-import { KeymapDropdown } from "./view/keymap-dropdown";
+import { ShortcutsList } from "./view/shortcuts-list";
 
 interface Preferences {
   delay: string;
@@ -27,7 +15,6 @@ export default function AppShortcuts(props?: { app: Application }) {
   const [keymaps, setKeymaps] = useState<string[]>([]);
   const [keymapSections, setKeymapSections] = useState<Section[]>([]);
   const [isLoading, setIsLoading] = useState(true);
-  const keyCodesResponse = useKeyCodes();
   const shortcutsProviderResponse = useAllShortcuts({ execute: !props?.app });
 
   useEffect(() => {
@@ -66,57 +53,14 @@ export default function AppShortcuts(props?: { app: Application }) {
     setKeymapSections(selectKeymap(application?.keymaps ?? [], newValue)?.sections ?? []);
   };
 
-  async function executeShortcut(bundleId: string | undefined, shortcutSequence: AtomicShortcut[]) {
-    if (keyCodesResponse.data === undefined) return;
-    const delay: number = parseFloat(getPreferenceValues<Preferences>().delay);
-    await closeMainWindow({ popToRootType: PopToRootType.Immediate });
-    await runShortcuts(bundleId, delay, shortcutSequence, keyCodesResponse.data);
-  }
-
   return (
-    <List
+    <ShortcutsList
       isLoading={isLoading}
-      searchBarPlaceholder="Search for shortcuts"
-      searchBarAccessory={<KeymapDropdown keymaps={keymaps} onKeymapChange={onKeymapChange} />}
-      navigationTitle={application?.name}
-    >
-      {keymapSections.map((section) => {
-        return (
-          <List.Section key={section.title} title={section.title}>
-            {section.hotkeys.map((shortcut) => {
-              return (
-                <List.Item
-                  key={shortcut.title}
-                  title={shortcut.title}
-                  subtitle={generateHotkeyText(shortcut)}
-                  accessories={
-                    shortcut.comment
-                      ? [
-                          {
-                            text: shortcut.comment,
-                            icon: Icon.SpeechBubble,
-                          },
-                        ]
-                      : undefined
-                  }
-                  keywords={[section.title]}
-                  actions={
-                    shortcut.sequence.length > 0 ? (
-                      <ActionPanel>
-                        <Action
-                          title="Apply"
-                          onAction={() => application && executeShortcut(application.bundleId, shortcut.sequence)}
-                        />
-                      </ActionPanel>
-                    ) : undefined
-                  }
-                />
-              );
-            })}
-          </List.Section>
-        );
-      })}
-    </List>
+      application={application}
+      keymaps={keymaps}
+      onKeymapChange={onKeymapChange}
+      keymapSections={keymapSections}
+    />
   );
 }
